test: 다리 길이, 다리 방향 입력값 유효성 검증 테스트 병합
