+++ __tests__/ValidationInputTest.js
@@ -15,3 +15,19 @@ describe("다리 길이 입력값 유효성 검증 테스트", () => {
     expect(result).toThrow();
   });
 });
+
+describe("다리 방향 입력값 유효성 검증 테스트", () => {
+  it("U, D가 아닌 값이 입력될 경우 예외처리", () => {
+    const inputDirection = "T";
+    const result = () => InputView.readMoving(inputDirection);
+
+    expect(result).toThrow();
+  });
+
+  it("공백이 입력될 경우 예외처리", () => {
+    const inputDirection = " ";
+    const result = () => InputView.readMoving(inputDirection);
+
+    expect(result).toThrow();
+  });
+});
